image-delete-and-update

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const { validationResult } = require('express-validator');
 
 const Product = require('../models/product');
+const fileHelper = require('../util/file')
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -156,6 +157,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl)
         product.imageUrl = image.path;
       }
       return product.save().then(result => {
@@ -191,8 +193,13 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
+  Product.findById(prodId).then(product => {
+    if (!product) {
+      return next(new Error('product not found'))
+    }
+    fileHelper.deleteFile(product.imageUrl)
+    return   Product.deleteOne({ _id: prodId, userId: req.user._id })
+  }).then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     })
